pyproject.toml
[tool.poetry]
name = "supa-workflows"
version = "0.1.0"
description = "Workflows written in Python, supported by Celery and RabbitMQ"
authors = ["devanshu <devanshukumar45@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
supabase = "^2.4.0"
pika = "^1.3.2"
celery = "^5.3.6"
python-dotenv = "^1.0.1"
gevent = "^24.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
